[SW-57] refactor: JS 내용 JSX로 최적화
